huge code base change - use of Config model
The before code base was so cluttered that I had to use a model here to save developer configuration instead of saving it and passing it one by one. The most tedious task by an interpreter is passing parameter between constructors and this was a bad code example to pass so many of them to StorageChooserBuilder constructor.

Implementation of config is fully tested and merged along with an option to use pass memory threshold and some usefull constants and pre-programmed preference key is also added. Memory threshold will be explained in the wiki seperately.