Only accept volumes which are directories
On some devices one of the volumes causes a failure when
its storage size is determined with setMemoryTotalSize().
The reason is that the volume could not stat with StatFs.doStat().

It is not clear which volumes are being attempted or why
they cannot be stat, as only the backtrace of the failures
are known. Expecting that the volume in question was not a directory,
requiring that all volumes be directories may avoid the failure.